% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concoct_directory.R
\name{concoct_directory}
\alias{concoct_directory}
\title{Get a directory of all nodes, files, and folders within a parent node}
\usage{
concoct_directory(id)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) (just "XXXX") of the parent node}
}
\value{
A vector with parent node information and options to look at component or file information using extensions e.g. origin$files$ComponentName will return an rjson style vector of all data on files within ComponentName whereas origin$components$ComponentName will return information on all components within ComponentName
}
\description{
This function is recursive and can take a long time. Please plan accordingly.
}
