% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all.R
\name{download_all}
\alias{download_all}
\title{Download all data files from each project (in "Data" folder) in main OSF project.}
\usage{
download_all(id = "judwb", path = getwd())
}
\arguments{
\item{id}{OSF id. Defaults to main Ag project.}

\item{path}{Path to download files. Defaults to working directory.}
}
\description{
Download all data files from each project (in "Data" folder) in main OSF project.
}
