% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_admin_sheet.R
\name{get_admin_sheet}
\alias{get_admin_sheet}
\title{Retreive Master Admin Project Sheet}
\usage{
get_admin_sheet(osf_file_id = "z8kpa")
}
\arguments{
\item{osf_file_id}{A string containing the code at the end of an OSF url.
For example, https://osf.io/z8kpa/.
The default string value is directly to the master sheet.}
}
\value{
A tibble in long format
}
\description{
Retreive Master Admin Project Sheet
}
