% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_all_file_links}
\alias{get_all_file_links}
\title{Get all file links from project}
\usage{
get_all_file_links(id = "judwb")
}
\arguments{
\item{id}{OSF id. Defaults to main Ag project.}
}
\description{
Get all file links from project
}
