% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get a data frame of csv files within a project component or folder}
\usage{
get_data(id = NULL, api_link = NULL)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) (just "XXXX") of the project node}

\item{api_link}{WaterButler link to use in place of id if no id exists}
}
\value{
A dataframe object
}
\description{
This function uses a project component id or WaterButler api link to return a dataframe of all csv files within the data folder of the project.
}
