% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Get a dictionary (vector) of files, folders, or children using the parent node id}
\usage{
get_dictionary(id, type)
}
\arguments{
\item{id}{OSF id (osf.io/XXXX) (just "XXXX") of the parent node}

\item{type}{Type of OSF child ("children", "files", "folders") for which you are obtaining a dictionary}
}
\value{
A dictionary (vector) with values = ids or WaterButler Links if no id available and keys = names of children
}
\description{
Get a dictionary (vector) of files, folders, or children using the parent node id
}
