% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_waterbutler.R
\name{get_node_url}
\alias{get_node_url}
\title{Get url at the end of a file path. Helper function for move_waterbutler()}
\usage{
get_node_url(list)
}
\arguments{
\item{list}{A vector of a file path}
}
\description{
Get url at the end of a file path. Helper function for move_waterbutler()
}
