% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_to_osf.R
\name{login_to_osf}
\alias{login_to_osf}
\title{Login to OSF (Open Science Framework)}
\usage{
login_to_osf(osf_pat = readr::read_lines("www/PAT"))
}
\arguments{
\item{osf_pat}{A Personal Access Token to access a private osf account.
This defaults to using a PAT stored as text file in www folder.
A string can be passed in here to sub for a different PAT.}
}
\value{
NULL - Only prints a statement welcoming user to OSF
}
\description{
Login to OSF (Open Science Framework)
}
