% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_waterbutler.R
\name{move_waterbutler}
\alias{move_waterbutler}
\title{Move a file that is untouched from one component or folder to another component or folder}
\usage{
move_waterbutler(from, to)
}
\arguments{
\item{from}{A file path beginning with the deepest level component id to the name of the file itself ("XXXX/foldername/filename")}

\item{from}{A file path beginning with the deepest level component id to the name of the destination folder itself ("XXXX/foldername/subfoldername") or simply a component id}
}
\description{
Move a file that is untouched from one component or folder to another component or folder
}
